LayoutTests/imported/w3c/web-platform-tests/wasm/serialization/module/serialization-via-history-expected.txt


PASS history.pushState(): simple case
PASS history.pushState(): is interleaved correctly
PASS history.replaceState(): simple case
PASS history.replaceState(): is interleaved correctly
